pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-mongodb>=0.11.0",
    "langchain-ollama>=1.0.1",
    "langchain-text-splitters>=1.1.1",
    "pymupdf>=1.27.1",
]
